Minor language clarification in SetupDraw example

+++ content/examples/Basics/Structure/SetupDraw/SetupDraw.pde
@@ -5,7 +5,7 @@
  * from top to bottom until the program is stopped.  
  */
 
-float y = 100;
+int y = 100;
 
 // The statements in the setup() function 
 // execute once when the program begins
@@ -19,7 +19,7 @@ void setup() {
 // sequence and after the last line is read, the first 
 // line is executed again.
 void draw() { 
-  background(0);   // Set the background to black
+  background(0);   // Clear the screen with a black background
   y = y - 1; 
   if (y < 0) { 
     y = height;
